backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-multipart
pandas
passlib[bcrypt]
python-jose[cryptography]
sentence-transformers
pgvector
groq
psycopg2-binary
tqdm
duckduckgo-search
